create MediaType package
